rpc: Refactoring the RPC calls
So far we always had the net/rpc style methods, in which methods would
get a request and a response reference and return only an error. This
feels a bit strange, so I am splitting the lightning RPC from the code
that is registered in the jsonrpc2 module to handle incoming
requests. `lightningrpc` now holds the pure RPC calls we can use from
other packages as well, while `webui/lightning.go` holds the shims to
expose these methods through the jsonrpc2 module.